Separate model from views
Since that point, all changes to the workflow should happen to the model. Both Editor and Canvas are now smart views who’s purpose is to render the model in a particular form and translate user interactions to model API calls.

Since our model is still somewhat dumb, it still requires editor to implement some part of business logic for everything to work, but Canvas is not required anymore and you can create a relatively simple stub to replace Ace editor with.