Make node creation syncronious
With virtualeditor in place, we can be certain that every operation on the model is syncronious. Previously, we returned a promise on creating a new node so the canvas would switch the node to edit mode after it gets rendered. It worked mosly by accident, due to nuances of Promise implementation. Now, Canvas will handle this on its own and ensure the component went through the whole lifecycle before we attempt to edit the node.