Fix rename after task on canvas
In previous commit that fixed switching between workflow specs an issue
was introduced that would not let you change any metadata under the
right sequence of events. Because the original runner_type was
overwritten on initial metadata write if you had a task on the canvas it
would trigger as though the runner was updated. This would put you into
a loop of not being able to update the metadata until you cleared the
canvas of all tasks. This commit fixes that issue by setting a proper
initial metadata and setting it to orquesta spec. It also moves the
check for changing spec inside the portion of metadata reducer that
handles spec setting on change.