Fixing mistral variables
Mistral variables are a pure object, while orquesta variables are a list
of objects with a single key. The original var functionality was built
around orquesta and assumed mistral handled vars the same way. This
commit fixes the incorrect handling of mistral vars by parsing the list
of objects into a flat object and visa-versa for loading.