Don't crash on mistral workbook
Despite not currently support mistral workbooks flow should not crash
when a workbook is loaded. This commit adds a check for workbooks and if
it is a workbook we emit a message to notify the user that we do not
currently support workbook. It also adds minor changes to the path
rendering to ensure that the workflow visualization can load.