Save new workflow twice
This commit fixes a bug in saving a new workflow after its initial save.
The bug was caused by the check done to see if the workflow is new or
not. When flow is loaded it downloads a list of actions. This list is
checked to see if the action is new or not. Once the initial save
happens the actions were not refreshed, thus the API call was always a
POST instead of an initial POST followed by a PUT. This commit refreshes
actions after the initial save.